requirements.txt
asgiref==3.3.4
boto3==1.17.107
botocore==1.20.107
Django==3.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
dnspython==2.1.0
flatten-dict==0.4.0
jmespath==0.10.0
Markdown==3.3.4
Pillow==8.2.0
Pygments==2.8.1
pymongo==3.11.3
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
